Moved auth.controller, auht.service, and auth.type to the facilities/auh directory.
