+++ src/controllers/messege.controller.ts
@@ -1,11 +1,11 @@
 import type { Request, Response } from "express";
+import { verifyUserStoreForOwnerAndManager } from "../features/common/authorization.service";
 import {
 	URI_CONNECT_LINE_GROUP,
 	URI_REGISTER_OWNER,
 	URI_REGISTER_STAFF,
 	URI_SHIFT_SUBMITTED,
 } from "../lib/env";
-import { verifyUserStoreForOwnerAndManager } from "../services/common/authorization.service";
 import {
 	sendGroupFlexMessage,
 	sendGroupMessageByTrigger,
