+++ src/features/auth/init/validation.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const storeIdValidate = z.object({
+	storeId: z.string({ message: "Invalid storeId format" }).uuid(),
+});
