+++ src/features/auth/lineAuth/type.ts
@@ -1,0 +1,13 @@
+export interface LineAuthServiceResponse {
+	userId: string;
+	name: string;
+	pictureUrl: string;
+}
+
+export interface LineAuthResponse {
+	ok: true;
+	userId: string;
+	name: string;
+	pictureUrl: string;
+	line_token: string;
+}
