+++ src/features/auth/reLogin/type.ts
@@ -1,0 +1,4 @@
+export interface ReLoginResponse {
+	ok: true;
+	user_token: string;
+}
