--- src/types/auth.type.ts
@@ -1,5 +1,0 @@
-export interface lineAuthResponse {
-	userId: string;
-	name: string;
-	pictureUrl: string;
-}
